+++ src/Provider/PaymentOption/CreditCardPaymentOptionProvider.php
@@ -214,6 +214,8 @@ class CreditCardPaymentOptionProvider implements PaymentOptionProviderInterface
         $paymentFeeData = $this->paymentFeeProvider->getPaymentFee($paymentMethod, $this->orderTotalProvider->getOrderTotal());
 
         if ($paymentFeeData->isActive()) {
+            $paymentFeeText = $this->paymentFeeProvider->getPaymentFeeText($paymentFeeData->getPaymentFeeTaxIncl());
+
             $paymentOption->setInputs(
                 array_merge($paymentOption->getInputs(), [
                     [
@@ -225,7 +227,7 @@ class CreditCardPaymentOptionProvider implements PaymentOptionProviderInterface
                         'type' => 'hidden',
                         'name' => 'payment-fee-price-display',
                         'value' => sprintf(
-                            $this->module->l('Payment Fee: %1s', self::FILE_NAME),
+                            $paymentFeeText,
                             $this->tools->displayPrice($paymentFeeData->getPaymentFeeTaxIncl())
                         ),
                     ],