+++ src/Provider/PaymentOption/CreditCardSingleClickPaymentOptionProvider.php
@@ -210,6 +210,8 @@ class CreditCardSingleClickPaymentOptionProvider implements PaymentOptionProvide
         $paymentFeeData = $this->paymentFeeProvider->getPaymentFee($paymentMethod, $this->orderTotalProvider->getOrderTotal());
 
         if ($paymentFeeData->isActive()) {
+            $paymentFeeText = $this->paymentFeeProvider->getPaymentFeeText($paymentFeeData->getPaymentFeeTaxIncl());
+
             $paymentOption->setInputs(
                 array_merge($paymentOption->getInputs(), [
                     [
@@ -221,7 +223,7 @@ class CreditCardSingleClickPaymentOptionProvider implements PaymentOptionProvide
                         'type' => 'hidden',
                         'name' => 'payment-fee-price-display',
                         'value' => sprintf(
-                            $this->module->l('Payment Fee: %1s', self::FILE_NAME),
+                            $paymentFeeText,
                             $this->tools->displayPrice($paymentFeeData->getPaymentFeeTaxIncl())
                         ),
                     ],