+++ changelog.md
@@ -1,6 +1,13 @@
 ![Mollie](https://www.mollie.nl/files/Mollie-Logo-Style-Small.png)
 
 # Changelog #
+
+## Changes in release 6.3.1 ##
++ Improved API key validation
++ Added Bizum payment method
++ Added Vipps payment method
++ Fixed minor issues with PrestaShop 9 compatibility
+
 ## Changes in release 6.3.0 ##
 + Apple Pay Direct fixed
 + Added DK, FI, PL, SE, NO, PT, CZ, SI, HU, ES translations